Move editorhub connection from App to EditorPage
